1 登录用户名长度限制为16位 2 账户管理用户名显示修改 3 增加设备，删除设备显示

+++ app/src/main/java/com/electric/cet/mobile/android/pq/Bean/DataBean.java
@@ -134,7 +134,7 @@ public class DataBean implements Serializable {
         return Longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         Longitude = longitude;
     }
 
@@ -142,7 +142,7 @@ public class DataBean implements Serializable {
         return Latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         Latitude = latitude;
     }
